Alterando os usos de Id de Integer para Long e refatorando fluxo de cartão e endereco

+++ src/main/java/br/com/fatec/ChopperHouseGames/core/repository/CartaoCreditoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartaoCreditoRepository extends JpaRepository<CartaoCredito, Integer> {
+public interface CartaoCreditoRepository extends JpaRepository<CartaoCredito, Long> {
 }
